+++ src/cmd/commands.hpp
@@ -77,6 +77,8 @@ class Commands {
 
   void findEmbseqs();
 
+  void findIncevalPrograms(const std::string& error_code);
+
   void lists();
 
   void compare(const std::string& path1, const std::string& path2);